Move the command to celery
