+++ pii-ner-exclude.txt
@@ -599,6 +599,7 @@ Updated Personal Email
 User
 User
 users.</td
+UUID
 value="update_status_form">
 Variables](/docs
 viewBox="0 0 483.2
